Proposal three steps: finished with idea
Finished up an idea of mine on how to show the three steps interface. Uses
text message bubbles to show the first two and uses a vague keyboard shape
for the third. This design emphasizes the steps by suggesting what each of them
would entail (text message and then computer device).

Added a new CSS atom for a one character disc. Added a new chatbox molecule
for the actual chat boxes.

Kept the computer design as inline as it may not be very reusable.